Refactor depth of market data updating
For DOM requests, IBKR sends an update command as:
    (side, position, add/update/delete, price, market)

Originally regular position-indexed lists were used to track the
updates, but sometimes the IBKR inbound updates would happen after
lists were cleared or other errors happened resulting in invalid position
indexes given.

To be more correct and reliable and accurate, we're moving to using
dicts instead of just postion-indexed lists for adding, updating,
and removing elements from an active DOM session.

Based on our network API debugging, the IBKR API always sends "insert"
requests sequentially before any update requests, so we can rely on
Python's built in dict insert-ordering behavior to maintain the order
of the DOM dict.

The performance impact of moving from lists to dicts is minimal
and also this code path only runs when live depth of market requests are
active.

We retain the same API usage where `ticker.domBids` and
`ticker.domAsks` are lists of price events in the correct order, but now we
also have `ticker.domBidsDict` and `ticker.domAsksDict` which are used
to _create_ the `domBids` and `domAsks` lists after each update, but the
user can access the underlying dicts directly too if useful.

This improvement also has an added benefit of isolating the `domBids`
and `domAsks` lists provided to users from the underlying data structure
we are using to create them (previously, if users directly modified the
`domBids` or `domAsks` lists, it could break the DOM session since all
the indexes for future live udpates would then be invalid).

Fixes #33