Add uninitialized dicts to the repr ignore checker
There's functionality to ignore unused/unset fields when printing
dataclasses here, but it wasn't checking for unused dicts, so we're
adding unused dict checking.

Previously the only container type used as defaults was lists, but now
we are using dicts in Ticker() for the depth of market logging, so we
are now ignoring empty dicts too from dataclass string representation
printing.