Implement automatic version metadata sync
This version value has historically been manually updated, but modern
python versions include a version library to read the version of the
library you are currently running, so the library itself can read its
own package version and provide it to users here.

Primary difference here is instead of manual version in tuple with
auto-string generation, now we use the package string then auto-generate
the tuple-of-integers version from the string version.

Fixes #47