Prevent an easy to trigger order modification bug
I've hit this error for a long time and it's probably better to place
the workaround directly as an API-level bug fix instead of everybody
needing to check for this bug on their own.

The problem is potentially our cached order objects let the
IBKR API overwrite fields we didn't set on our own (so when we go to
submit order updates on cached order objects, we submit data we never
set ourselves).

The fix for the IBKR API overwriting our own order fields is probably
to create a secondary "IBKR Updated Order Metadata" object attached
to our real orders so the API updates don't conflict with a user's
own exact order data, but that's more work and this quick fix solved
this order update problem for now (along with the error handler
validation check fixes).

There is one other big order placement bug where if you order a spread
using limit order adaptive fast/slow, the IBKR API changes the order
type to IBALGO, but you _can't_ submit order updates with IBAGLO as an
`orderType` yourself (I'm not sure if you can ever use it?) so also I
have other bug fixes where "orderType = 'LMT' if orderType == "IBALGO"
else orderType", but I'm not sure if that is also a universal problem in
need if fixing for _everybody_ yet.