Allow Decimal type checking as price/qty in Order
Just helper logic so mypy doesn't complain if we use Decimal values
directly on these instance variables.

It _is_ safe to use floats or Doubles for these values because when
these values are sent back to IBKR over the network, we just run
str(value) on all of them, so int/float/Decimal are all interchangeable
as far as the actual protocol API level is concerned (because the IBKR
gateway API protocol is 100% ASCII string based).