Stop guarding "last size/price" updates on values
For some reason the Ticker objects here have always maintained a
"previous value" entry, but it only updates the previous value entry if
the current value isn't the previous value.

But isn't that wrong? If you want the previous trade price, it should be
the true "previously seen trade price" and not "only the previous price
if it isn't the current price?"

Also, the historical logic was updating size and price independently, so
the "lastPrice" and "lastSize" details weren't guaranteed to even match
in time as to being one matches the other?

Now previous price is the previously received price and size data
regardless of if it matches the current price details or not.